# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/ext/add": {
      "post": {
        "tags": [
          "Extensions"
        ],
        "summary": "Add Extension",
        "operationId": "add_extension_ext_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Extension"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/ext/delete/{extension_id}": {
      "delete": {
        "tags": [
          "Extensions"
        ],
        "summary": "Delete Extension",
        "operationId": "delete_extension_ext_delete__extension_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "integer"
            },
            "name": "extension_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Id",
                "type": "integer",
                "description": "ID of the extension"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Extension"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/ext/update/{extension_id}": {
      "patch": {
        "tags": [
          "Extensions"
        ],
        "summary": "Update Extension",
        "description": "Update an extension. Only the owner can do this. Only non-null fields will be updated.",
        "operationId": "update_extension_ext_update__extension_id__patch",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "integer"
            },
            "name": "extension_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionUpdateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Extension"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/ext/list": {
      "get": {
        "tags": [
          "Extensions"
        ],
        "summary": "List Extensions",
        "operationId": "list_extensions_ext_list_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Skip",
              "type": "integer",
              "default": 0
            },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Extensions Ext List Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Extension"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ext/get/{extension_id}": {
      "get": {
        "tags": [
          "Extensions"
        ],
        "summary": "Get Extension",
        "operationId": "get_extension_ext_get__extension_id__get",
        "parameters": [
          {
            "description": "`id` (int) or `name` (string) of the extension",
            "required": true,
            "schema": {
              "title": "Extension Id",
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ],
              "description": "`id` (int) or `name` (string) of the extension"
            },
            "name": "extension_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Extension"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/version/add/{extension_id}": {
      "post": {
        "tags": [
          "Versions"
        ],
        "summary": "Add a new version to an extension",
        "operationId": "add_extension_version_version_add__extension_id__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "integer"
            },
            "name": "extension_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionVersionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionVersion"
                }
              }
            }
          },
          "409": {
            "description": "Version number already exists for this extension"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/version/delete/{version_id}": {
      "delete": {
        "tags": [
          "Versions"
        ],
        "summary": "Delete Extension Version",
        "operationId": "delete_extension_version_version_delete__version_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            },
            "name": "version_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Description",
              "default": "Delete an extension"
            },
            "name": "description",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionVersion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/version/update/{version_id}": {
      "patch": {
        "tags": [
          "Versions"
        ],
        "summary": "Update a version.",
        "description": "Only the extension owner can do this. Only non-null fields will be updated.",
        "operationId": "update_extension_version_version_update__version_id__patch",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            },
            "name": "version_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionVersionUpdateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionVersion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/version/download/{version_id}": {
      "get": {
        "tags": [
          "Versions"
        ],
        "summary": "Download a extension zip file for the given version",
        "operationId": "download_extension_version_file_version_download__version_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            },
            "name": "version_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/version/upload/{version_id}": {
      "post": {
        "tags": [
          "Versions"
        ],
        "summary": "Upload a extension zip file for a version.",
        "operationId": "upload_extension_version_file_version_upload__version_id__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            },
            "name": "version_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_extension_version_file_version_upload__version_id__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionFile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/version/list/{extension_id}": {
      "get": {
        "tags": [
          "Versions"
        ],
        "summary": "List all versions for an extension",
        "operationId": "list_extensions_version_list__extension_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Extension Id",
              "type": "integer"
            },
            "name": "extension_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Skip",
              "type": "integer",
              "default": 0
            },
            "name": "skip",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "type": "integer",
              "default": 100
            },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Extensions Version List  Extension Id  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExtensionVersion"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_extension_version_file_version_upload__version_id__post": {
        "title": "Body_upload_extension_version_file_version_upload__version_id__post",
        "required": [
          "file"
        ],
        "type": "object",
        "properties": {
          "file": {
            "title": "File",
            "type": "string",
            "format": "binary"
          }
        }
      },
...
      "Extension": {
        "title": "Extension",
        "required": [
          "name",
          "owner"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "Short name of the extension"
          },
          "description_short": {
            "title": "Description Short",
            "type": "string",
            "description": "A short description what the extension is about. Max characters: 140",
            "default": ""
          },
          "description_long": {
            "title": "Description Long",
            "type": "string",
            "description": "An elaborate description what the extension does",
            "default": ""
          },
          "id": {
            "title": "Id",
            "type": "integer",
            "description": "UUID of the extension"
          },
          "owner": {
            "title": "Owner",
            "type": "string",
            "description": "UUID of the extension owner"
          },
          "versions": {
            "title": "Versions",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtensionVersion"
            },
            "default": []
          }
        }
      },
      "ExtensionInput": {
        "title": "ExtensionInput",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "Short name of the extension"
          },
          "description_short": {
            "title": "Description Short",
            "type": "string",
            "description": "A short description what the extension is about. Max characters: 140",
            "default": ""
          },
          "description_long": {
            "title": "Description Long",
            "type": "string",
            "description": "An elaborate description what the extension does",
            "default": ""
          }
        }
      },
      "ExtensionUpdateInput": {
        "title": "ExtensionUpdateInput",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "Short name of the extension"
          },
          "description_short": {
            "title": "Description Short",
            "type": "string",
            "description": "A short description what the extension is about. Max characters: 140"
          },
          "description_long": {
            "title": "Description Long",
            "type": "string",
            "description": "An elaborate description what the extension does"
          }
        }
      },
      "ExtensionVersion": {
        "title": "ExtensionVersion",
        "required": [
          "major",
          "minor",
          "patch",
          "semver",
          "id",
          "extension_id",
          "is_latest"
        ],
        "type": "object",
        "properties": {
          "major": {
            "title": "Major",
            "type": "integer"
          },
          "minor": {
            "title": "Minor",
            "type": "integer"
          },
          "patch": {
            "title": "Patch",
            "type": "integer"
          },
          "note": {
            "title": "Note",
            "type": "string",
            "default": ""
          },
          "status": {
            "title": "Status",
            "type": "integer",
            "description": "\n        0: `draft` - For testing, not yet released, only visible to extension admins.\n        1: `private` - Released, but for visible testers only.\n        2: `public` - For public consumption.\n        3: `deprecated` - No longer recommended for use.\n        4: `withdrawn` - No longer available for use. Use a different version.\n        ",
            "default": 0
          },
          "semver": {
            "title": "Semver",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "extension_id": {
            "title": "Extension Id",
            "type": "integer"
          },
          "is_latest": {
            "title": "Is Latest",
            "type": "boolean"
          },
          "If this version has a file attached to it. A version must have an source file to publish it officially.": {
            "title": "If This Version Has A File Attached To It. A Version Must Have An Source File To Publish It Officially.",
            "type": "string"
          }
        }
      },
      "ExtensionVersionInput": {
        "title": "ExtensionVersionInput",
        "required": [
          "major",
          "minor",
          "patch"
        ],
        "type": "object",
        "properties": {
          "major": {
            "title": "Major",
            "type": "integer"
          },
          "minor": {
            "title": "Minor",
            "type": "integer"
          },
          "patch": {
            "title": "Patch",
            "type": "integer"
          },
          "note": {
            "title": "Note",
            "type": "string",
            "default": ""
          },
          "status": {
            "title": "Status",
            "type": "integer",
            "description": "\n        0: `draft` - For testing, not yet released, only visible to extension admins.\n        1: `private` - Released, but for visible testers only.\n        2: `public` - For public consumption.\n        3: `deprecated` - No longer recommended for use.\n        4: `withdrawn` - No longer available for use. Use a different version.\n        ",
            "default": 0
          }
        }
      },
      "ExtensionVersionUpdateInput": {
        "title": "ExtensionVersionUpdateInput",
        "required": [
          "major",
          "minor",
          "patch",
          "note"
        ],
        "type": "object",
        "properties": {
          "major": {
            "title": "Major",
            "type": "integer"
          },
          "minor": {
            "title": "Minor",
            "type": "integer"
          },
          "patch": {
            "title": "Patch",
            "type": "integer"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer",
            "description": "\n        0: `draft` - For testing, not yet released, only visible to extension admins.\n        1: `private` - Released, but for visible testers only.\n        2: `public` - For public consumption.\n        3: `deprecated` - No longer recommended for use.\n        4: `withdrawn` - No longer available for use. Use a different version.\n        ",
            "default": 0
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
      "VersionFile": {
        "title": "VersionFile",
        "required": [
          "id",
          "file_name",
          "version_id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "description": "UUID of the extension"
          },
          "file_name": {
            "title": "File Name",
            "type": "string",
            "description": "Original filename"
          },
          "version_id": {
            "title": "Version Id",
            "type": "integer",
            "description": "ID of the version this file belongs to"
          }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fastapiusersauth"
      }
    }
  }
}